requirements.txt
pylint
flask
pycrypto
statistics
runstats
coverage
nose
pycryptodome
pkcs7
pytest